+++ access-token-manager/pom.xml
@@ -45,8 +45,8 @@
     <properties>
         <junit.version>4.11</junit.version>
         <mockito.version>1.9.5</mockito.version>
-        <spring.version>4.0.1.RELEASE</spring.version>
-        <jackson.version>2.3.0</jackson.version>
+        <spring.version>4.0.3.RELEASE</spring.version>
+        <jackson.version>2.3.3</jackson.version>
         <apache.commons.cli.version>1.2</apache.commons.cli.version>
         <slf4j.version>1.7.5</slf4j.version>
         <logback.version>1.0.13</logback.version>
