+++ admin-console/pom.xml
@@ -176,7 +176,7 @@
         <dependency>
             <groupId>org.aspectj</groupId>
             <artifactId>aspectjrt</artifactId>
-            <version>1.7.2</version>
+            <version>1.8.0</version>
         </dependency>
 
         <!-- Logging -->
