+++ crafter-security-provider/src/main/java/org/craftercms/security/exception/AuthenticationRequiredException.java
@@ -23,6 +23,10 @@ package org.craftercms.security.exception;
  */
 public class AuthenticationRequiredException extends AuthenticationException {
 
+    public AuthenticationRequiredException(String s) {
+        super(s);
+    }
+
     public AuthenticationRequiredException(String s, Throwable throwable) {
         super(s, throwable);
     }
