+++ crafter-security-provider/src/main/java/org/craftercms/security/processors/RequestSecurityProcessorChain.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.craftercms.security.api;
+package org.craftercms.security.processors;
 
 import org.craftercms.commons.http.RequestContext;
