--- crafter-security-provider/src/test/java/org/craftercms/security/testing/unit/TestPlaceholder.java
@@ -1,10 +1,0 @@
-package org.craftercms.security.testing.unit;
-
-import org.junit.Test;
-
-public class TestPlaceholder {
-    @Test
-    public void Ok() {
-
-    }
-}
