+++ integration-tests/pom.xml
@@ -19,7 +19,7 @@
     <properties>
         <junit.version>4.11</junit.version>
         <mockito.version>1.9.5</mockito.version>
-        <spring.version>4.0.1.RELEASE</spring.version>
+        <spring.version>4.0.3.RELEASE</spring.version>
     </properties>
 
     <build>
