+++ server/pom.xml
@@ -44,9 +44,9 @@
         <jetty.version>8.1.2.v20120308</jetty.version>
 
         <java.version>1.7</java.version>
-        <org.springframework.version>4.0.1.RELEASE</org.springframework.version>
+        <org.springframework.version>4.0.3.RELEASE</org.springframework.version>
         <org.spring-security.version>3.2.0.RELEASE</org.spring-security.version>
-        <org.aspectj.version>1.7.2</org.aspectj.version>
+        <org.aspectj.version>1.8.0</org.aspectj.version>
         <cglib.version>2.2</cglib.version>
         <org.slf4j.version>1.7.5</org.slf4j.version>
         <javax.inject.version>1</javax.inject.version>
@@ -60,7 +60,7 @@
         <org.apache.commons.fileupload.version>1.3</org.apache.commons.fileupload.version>
         <org.apache.commons.io.version>2.4</org.apache.commons.io.version>
         <org.apache.commons.codec.version>1.9</org.apache.commons.codec.version>
-        <com.fasterxml.jackson.version>2.3.0</com.fasterxml.jackson.version>
+        <com.fasterxml.jackson.version>2.3.3</com.fasterxml.jackson.version>
         <com.thoughtworks.xstream.version>1.4.2</com.thoughtworks.xstream.version>
         <org.mongodb.java-driver.version>2.11.3</org.mongodb.java-driver.version>
