+++ server/src/main/resources/crafter/profile/init-data.js
@@ -1,15 +1,39 @@
 if (db.accesstoken.count() == 0) {
-	db.accesstoken.insert({
-		"_id" : "e8f5170c-877b-416f-b70f-4b09772f8e2d",
-		"application" : "adminconsole",
-		"tenantPermissions" : [
-			{
-                "allowedActions" : [ "*" ],
-				"tenant" : "*"
-			}
-		],
-		"expiresOn" : new Date("Jan 1, 2024")
-	});
+	db.accesstoken.insert([
+        {
+            "_id" : "e8f5170c-877b-416f-b70f-4b09772f8e2d",
+            "application" : "adminconsole",
+            "tenantPermissions" : [
+                {
+                    "allowedActions" : [ "*" ],
+                    "tenant" : "*"
+                }
+            ],
+            "expiresOn" : new Date("Jan 1, 2024")
+	    },
+        {
+            "_id" : "b4d44030-d0af-11e3-9c1a-0800200c9a66",
+            "application" : "crafterengine",
+            "tenantPermissions" : [
+                {
+                    "allowedActions" : [ "*" ],
+                    "tenant" : "*"
+                }
+            ],
+            "expiresOn" : new Date("Jan 1, 2024")
+        },
+        {
+            "_id" : "2ba3ac10-c43e-11e3-9c1a-0800200c9a66",
+            "application" : "craftersocial",
+            "tenantPermissions" : [
+                {
+                    "allowedActions" : [ "*" ],
+                    "tenant" : "*"
+                }
+            ],
+            "expiresOn" : new Date("Jan 1, 2024")
+        }
+    ]);
 }
 if (db.tenant.count() == 0) {
     db.tenant.insert({
